llm/client_server_openai/pyproject.toml
[tool.poetry]
name = "gradio_openai"
version = "0.1.0"
description = ""
authors = ["Vicente Herrera <vicenteherrera@vicenteherrera.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

# [[tool.poetry.source]]
# name = "devpi"
# url = "http://127.0.0.1:3141/root/pypi/+simple/"
# priority = "primary"
# Disable PyPI and uncomment this to use devpi local library cache

[tool.poetry.dependencies]
python = "^3.11.7"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
langchain = "^0.2.1"
jsonify = "^0.5"
langchain-community = "^0.2.1"
flask = "^3.0.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"
